+++ src/hooks/useSettings.ts
@@ -25,6 +25,15 @@ const initialState = mergeDeepRight(
   window.leckerliSettings ?? {}
 );
 
+// Define cookie configuration
+const cookieConfig: {
+  sameSite: 'strict';
+  domain: string;
+} = {
+  sameSite: 'strict',
+  domain: initialState.domain,
+};
+
 // Get cookie value
 const initialCookie = cookies.get(initialState.name);
 
@@ -49,9 +58,11 @@ const useSettings = create<SettingsStore>((set, getState) => ({
   // Write cookie if not exist and setup eventListeners
   init: () => {
     if (isNil(initialCookie) && getState().choiceMade) {
-      cookies.set(getState().name, JSON.stringify(getState().cookie), {
-        domain: window.location.host,
-      });
+      cookies.set(
+        getState().name,
+        JSON.stringify(getState().cookie),
+        cookieConfig
+      );
     }
 
     // Emit initial event and data or null if the choice has not been made
@@ -105,9 +116,7 @@ const useSettings = create<SettingsStore>((set, getState) => ({
         return acc;
       }, state.baseData);
 
-      cookies.set(state.name, JSON.stringify(newCookie), {
-        domain: window.location.host,
-      });
+      cookies.set(state.name, JSON.stringify(newCookie), cookieConfig);
 
       // Emit event and data
       document.dispatchEvent(
@@ -133,9 +142,7 @@ const useSettings = create<SettingsStore>((set, getState) => ({
         return acc;
       }, state.baseData);
 
-      cookies.set(state.name, JSON.stringify(newCookie), {
-        domain: window.location.host,
-      });
+      cookies.set(state.name, JSON.stringify(newCookie), cookieConfig);
 
       // Emit event and data
       document.dispatchEvent(